+++ src/components/EChartsPanel/EChartsPanel.test.tsx
