+++ src/components/EChartsPanel/EChartsPanel.tsx
